split files for fm strategies and i_rebalancer
+ simplification in fm strategy
