+++ mt-kahypar/partition/refinement/fm/multitry_kway_fm.h
@@ -32,6 +32,7 @@
 #include "mt-kahypar/partition/context.h"
 
 #include "mt-kahypar/partition/refinement/i_refiner.h"
+#include "mt-kahypar/partition/refinement/i_rebalancer.h"
 #include "mt-kahypar/partition/refinement/fm/localized_kway_fm_core.h"
 #include "mt-kahypar/partition/refinement/fm/global_rollback.h"
 #include "mt-kahypar/partition/refinement/fm/strategies/i_fm_strategy.h"
