+++ mt-kahypar/partition/refinement/fm/strategies/gain_cache_strategy.h
@@ -3,7 +3,7 @@
  *
  * This file is part of Mt-KaHyPar.
  *
- * Copyright (C) 2020 Lars Gottesbüren <lars.gottesbueren@kit.edu>
+ * Copyright (C) 2023 Nikolai Maas <nikolai.maas@kit.edu>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,263 +26,12 @@
 
 #pragma once
 
-#include "mt-kahypar/partition/refinement/fm/fm_commons.h"
 #include "mt-kahypar/partition/refinement/fm/localized_kway_fm_core.h"
 #include "mt-kahypar/partition/refinement/fm/strategies/i_fm_strategy.h"
+#include "mt-kahypar/partition/refinement/fm/strategies/local_gain_cache_strategy.h"
 
 
 namespace mt_kahypar {
-
-  /*
-   * LocalFMStrategy interface
-   * static constexpr bool uses_gain_cache
-   * static constexpr bool maintain_gain_cache_between_rounds
-   * static constexpr bool is_unconstrained
-   *
-   * Constructor(context, sharedData, blockPQ, vertexPQs, runStats)
-   * insertIntoPQ(phg, gain_cache, node)
-   * updateGain(phg, gain_cache, node, move)
-   * findNextMove(phg, gain_cache, move)
-   * applyMove(phg, gain_cache, move, global)
-   * reset()
-   * deltaGainUpdates(phg, gain_cache, sync_update)
-   *
-   */
-
-class LocalGainCacheStrategy {
-public:
-
-  using BlockPriorityQueue = ds::ExclusiveHandleHeap< ds::MaxHeap<Gain, PartitionID> >;
-  using VertexPriorityQueue = ds::MaxHeap<Gain, HypernodeID>;    // these need external handles
-
-  static constexpr bool uses_gain_cache = true;
-  static constexpr bool maintain_gain_cache_between_rounds = true;
-  static constexpr bool is_unconstrained = false;
-
-  LocalGainCacheStrategy(const Context& context,
-                         FMSharedData& sharedData,
-                         BlockPriorityQueue& blockPQ,
-                         vec<VertexPriorityQueue>& vertexPQs,
-                         FMStats& runStats) :
-      context(context),
-      runStats(runStats),
-      sharedData(sharedData),
-      blockPQ(blockPQ),
-      vertexPQs(vertexPQs) { }
-
-  template<typename PartitionedHypergraph, typename GainCache>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  void insertIntoPQ(const PartitionedHypergraph& phg,
-                    const GainCache& gain_cache,
-                    const HypernodeID v) {
-    const PartitionID pv = phg.partID(v);
-    ASSERT(pv < context.partition.k);
-    auto [target, gain] = computeBestTargetBlock(phg, gain_cache, v, pv);
-    ASSERT(target < context.partition.k, V(target) << V(context.partition.k));
-    sharedData.targetPart[v] = target;
-    vertexPQs[pv].insert(v, gain);  // blockPQ updates are done later, collectively.
-    runStats.pushes++;
-  }
-
-  template<typename PartitionedHypergraph, typename GainCache>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  void updateGain(const PartitionedHypergraph& phg,
-                  const GainCache& gain_cache,
-                  const HypernodeID v,
-                  const Move& move) {
-    const PartitionID pv = phg.partID(v);
-    ASSERT(vertexPQs[pv].contains(v));
-    const PartitionID designatedTargetV = sharedData.targetPart[v];
-    Gain gain = 0;
-    PartitionID newTarget = kInvalidPartition;
-
-    if (context.partition.k < 4 || designatedTargetV == move.from || designatedTargetV == move.to) {
-      // penalty term of designatedTargetV is affected.
-      // and may now be greater than that of other blocks --> recompute full
-      std::tie(newTarget, gain) = computeBestTargetBlock(phg, gain_cache, v, pv);
-    } else {
-      // penalty term of designatedTargetV is not affected.
-      // only move.from and move.to may be better
-      std::tie(newTarget, gain) = bestOfThree(phg, gain_cache,
-        v, pv, { designatedTargetV, move.from, move.to });
-    }
-
-    sharedData.targetPart[v] = newTarget;
-    vertexPQs[pv].adjustKey(v, gain);
-  }
-
-  template<typename PartitionedHypergraph, typename GainCache>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  bool findNextMove(const PartitionedHypergraph& phg,
-                    const GainCache& gain_cache,
-                    Move& m) {
-    updatePQs();
-
-    if (blockPQ.empty()) {
-      return false;
-    }
-
-    while (true) {
-      const PartitionID from = blockPQ.top();
-      const HypernodeID u = vertexPQs[from].top();
-      const Gain estimated_gain = vertexPQs[from].topKey();
-      ASSERT(estimated_gain == blockPQ.topKey());
-      auto [to, gain] = computeBestTargetBlock(phg, gain_cache, u, phg.partID(u));
-
-      if (gain >= estimated_gain) { // accept any gain that is at least as good
-        m.node = u; m.to = to; m.from = from;
-        m.gain = gain;
-        runStats.extractions++;
-        vertexPQs[from].deleteTop();  // blockPQ updates are done later, collectively.
-        return true;
-      } else {
-        runStats.retries++;
-        vertexPQs[from].adjustKey(u, gain);
-        sharedData.targetPart[u] = to;
-        if (vertexPQs[from].topKey() != blockPQ.keyOf(from)) {
-          blockPQ.adjustKey(from, vertexPQs[from].topKey());
-        }
-      }
-    }
-  }
-
-  template<typename PartitionedHypergraph, typename GainCache>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  void applyMove(const PartitionedHypergraph&, const GainCache&, Move, bool) {
-    // nothing to do here
-  }
-
-  template<typename PartitionedHypergraph, typename GainCache>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  void revertMove(const PartitionedHypergraph&, const GainCache&, Move, bool) {
-    // nothing to do here
-  }
-
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  void flushLocalChanges() {
-    // nothing to do here
-  }
-
-  void reset() {
-    // release all nodes that were not moved
-    const bool release = sharedData.release_nodes
-                         && runStats.moves > 0;
-
-    if (release) {
-      // Release all nodes contained in PQ
-      for (PartitionID i = 0; i < context.partition.k; ++i) {
-        for (PosT j = 0; j < vertexPQs[i].size(); ++j) {
-          const HypernodeID v = vertexPQs[i].at(j);
-          sharedData.nodeTracker.releaseNode(v);
-        }
-      }
-    }
-
-    for (PartitionID i = 0; i < context.partition.k; ++i) {
-      vertexPQs[i].clear();
-    }
-    blockPQ.clear();
-  }
-
-
-  // We're letting the FM details implementation decide what happens here, since some may not want to do gain cache updates,
-  // but rather update gains in their PQs or something
-
-  template<typename PartitionedHypergraph, typename GainCache>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  void deltaGainUpdates(PartitionedHypergraph& phg,
-                        GainCache& gain_cache,
-                        const SynchronizedEdgeUpdate& sync_update) {
-    gain_cache.deltaGainUpdate(phg, sync_update);
-  }
-
-private:
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  void updatePQs() {
-    for (PartitionID i = 0; i < context.partition.k; ++i) {
-      if (!vertexPQs[i].empty()) {
-        blockPQ.insertOrAdjustKey(i, vertexPQs[i].topKey());
-      } else if (blockPQ.contains(i)) {
-        blockPQ.remove(i);
-      }
-    }
-  }
-
-  template<typename PartitionedHypergraph, typename GainCache>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  std::pair<PartitionID, HyperedgeWeight> computeBestTargetBlock(const PartitionedHypergraph& phg,
-                                                                 const GainCache& gain_cache,
-                                                                 const HypernodeID u,
-                                                                 const PartitionID from) {
-    const HypernodeWeight wu = phg.nodeWeight(u);
-    const HypernodeWeight from_weight = phg.partWeight(from);
-    PartitionID to = kInvalidPartition;
-    HyperedgeWeight to_benefit = std::numeric_limits<HyperedgeWeight>::min();
-    HypernodeWeight best_to_weight = from_weight - wu;
-    for ( const PartitionID& i : gain_cache.adjacentBlocks(u) ) {
-      if (i != from) {
-        const HypernodeWeight to_weight = phg.partWeight(i);
-        const HyperedgeWeight penalty = gain_cache.benefitTerm(u, i);
-        if ( ( penalty > to_benefit || ( penalty == to_benefit && to_weight < best_to_weight ) ) &&
-             to_weight + wu <= context.partition.max_part_weights[i] ) {
-          to_benefit = penalty;
-          to = i;
-          best_to_weight = to_weight;
-        }
-      }
-    }
-    const Gain gain = to != kInvalidPartition ? to_benefit - gain_cache.penaltyTerm(u, phg.partID(u))
-                                              : std::numeric_limits<HyperedgeWeight>::min();
-    return std::make_pair(to, gain);
-  }
-
-  template<typename PartitionedHypergraph, typename GainCache>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE
-  std::pair<PartitionID, HyperedgeWeight> bestOfThree(const PartitionedHypergraph& phg,
-                                                      const GainCache& gain_cache,
-                                                      HypernodeID u,
-                                                      PartitionID from,
-                                                      std::array<PartitionID, 3> parts) {
-
-    const HypernodeWeight wu = phg.nodeWeight(u);
-    const HypernodeWeight from_weight = phg.partWeight(from);
-    PartitionID to = kInvalidPartition;
-    HyperedgeWeight to_benefit = std::numeric_limits<HyperedgeWeight>::min();
-    HypernodeWeight best_to_weight = from_weight - wu;
-    for (PartitionID i : parts) {
-      if (i != from && i != kInvalidPartition) {
-        const HypernodeWeight to_weight = phg.partWeight(i);
-        const HyperedgeWeight penalty = gain_cache.benefitTerm(u, i);
-        if ( ( penalty > to_benefit || ( penalty == to_benefit && to_weight < best_to_weight ) ) &&
-             to_weight + wu <= context.partition.max_part_weights[i] ) {
-          to_benefit = penalty;
-          to = i;
-          best_to_weight = to_weight;
-        }
-      }
-    }
-    const Gain gain = to != kInvalidPartition ? to_benefit - gain_cache.penaltyTerm(u, phg.partID(u))
-                                              : std::numeric_limits<HyperedgeWeight>::min();
-    return std::make_pair(to, gain);
-  }
-
-  const Context& context;
-
-  FMStats& runStats;
-
-protected:
-  FMSharedData& sharedData;
-
-  // ! Priority Queue that contains for each block of the partition
-  // ! the vertex with the best gain value
-  BlockPriorityQueue& blockPQ;
-
-  // ! From PQs -> For each block it contains the vertices (contained
-  // ! in that block) touched by the current local search associated
-  // ! with their gain values
-  vec<VertexPriorityQueue>& vertexPQs;
-};
-
 
 template<typename TypeTraits, typename GainTypes>
 class GainCacheStrategy: public IFMStrategy {
